+++ src/Telegram/Types/Inline/InlineQueryResultPhoto.php
@@ -139,7 +139,7 @@ class InlineQueryResultPhoto extends InlineQueryResult
         ?string $title = null,
         ?string $description = null,
         ?string $caption = null,
-        ?ParseMode $parse_mode = null,
+        ParseMode|string|null $parse_mode = null,
         ?array $caption_entities = null,
         ?InlineKeyboardMarkup $reply_markup = null,
         ?InputMessageContent $input_message_content = null,
@@ -165,7 +165,7 @@ class InlineQueryResultPhoto extends InlineQueryResult
     public function jsonSerialize(): array
     {
         return array_filter_null([
-            'type' => $this->type->value,
+            'type' => $this->type,
             'id' => $this->id,
             'photo_url' => $this->photo_url,
             'thumb_url' => $this->thumbnail_url,
@@ -174,7 +174,7 @@ class InlineQueryResultPhoto extends InlineQueryResult
             'title' => $this->title,
             'description' => $this->description,
             'caption' => $this->caption,
-            'parse_mode' => $this->parse_mode?->value,
+            'parse_mode' => $this->parse_mode,
             'caption_entities' => $this->caption_entities,
             'reply_markup' => $this->reply_markup,
             'input_message_content' => $this->input_message_content,